Actualizado 100% Funcional

+++ app/src/main/java/com/nexohogar/domain/model/Transaction.kt
@@ -2,12 +2,12 @@ package com.nexohogar.domain.model
 
 /**
  * Modelo de dominio para una Transacción.
- * Representa la entidad de negocio simplificada.
+ * amount en Long porque CLP no tiene decimales.
  */
 data class Transaction(
     val id: String,
     val description: String?,
-    val amount: Double,
+    val amount: Long,           // Long: CLP no tiene decimales
     val accountId: String,
     val createdAt: String,
     val type: String
